uv3p: Split verify_run repository tests under the size guard
PR CI's Server Size Guard flagged
server/crates/djinn-db/src/repositories/verify_run.rs (52,479 bytes >
51,200) after the cache-lookup addition. Convert it to a directory module
following the existing repositories/note/ precedent: code in
verify_run/mod.rs, the test module in verify_run/tests.rs
(#[cfg(test)] mod tests;). Pure code motion, no behavior change.

The dedent of the moved test module changed whitespace inside its
multi-line sqlx::query! literals, so the .sqlx offline cache was
regenerated (one new entry for the re-hashed test query).
