+++ server/.sqlx/query-039360567d383d942753b84f546d6d1ea90ce2780309cff1803a508ae00e36d7.json
@@ -1,0 +1,17 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO tasks (id, project_id, short_id, epic_id, title, description, design,\n                            issue_type, priority, owner, status, continuation_count, labels, acceptance_criteria, memory_refs)\n         VALUES ($1, $2, $3, $4, 'Task', '', '', 'task', 0, '', 'open', 0, '[]'::jsonb, '[]'::jsonb, '[]'::jsonb)",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Varchar",
+        "Varchar",
+        "Varchar",
+        "Varchar"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "039360567d383d942753b84f546d6d1ea90ce2780309cff1803a508ae00e36d7"
+}
